feat: enhance modularity and behavior consistency across animal systems
- Added `BeeAccessor` and `FoxAccessor` for accessing private methods in vanilla classes.
- Improved interaction with sleeping and trust behaviors for foxes using mixins.
- Refactored `HiveDefenseBehavior` to support living entity targets.
- Updated `NestBuildingHandle` to leverage handle tags for better NBT management.
- Standardized positional calculations using `Vec3` across wolf and villager behaviors.
- Updated particle effects and sounds for sheep and wolves to match their context.
- Removed unused getters and setters to streamline codebase.
- Enhanced `SnowEatingBehavior` with deeper integration into ecology states.
- Updated `MooshroomMilkProductionHandle` to utilize data components for stew effects.

+++ src/main/java/me/javavirtualenv/behavior/feline/FallDamageReductionBehavior.java
@@ -2,7 +2,7 @@ package me.javavirtualenv.behavior.feline;
 
 import me.javavirtualenv.behavior.core.BehaviorContext;
 import me.javavirtualenv.behavior.core.Vec3d;
-import me.javavirtualenv.behavior.steering.SteeringBehavior;
+import me.javavirtualenv.behavior.core.SteeringBehavior;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.level.Level;
 import net.minecraft.sounds.SoundEvents;
@@ -88,7 +88,7 @@ public class FallDamageReductionBehavior extends SteeringBehavior {
                 mob.fallDistance = 0;
             } else {
                 // Reduce damage by 80%
-                mob.fallDistance = fallDistance * 0.2;
+                mob.fallDistance = (float)(fallDistance * 0.2);
             }
         }
     }
